+++ source/CCustomScript.cpp
@@ -148,6 +148,22 @@ CCustomScript::CCustomScript(const char* szFileName, bool bIsMiss, CRunningScrip
             is.read(reinterpret_cast<char*>(BaseIP), length);
 
             m_codeSize = length;
+
+            // check for extra SCM data at the end
+            static constexpr char SB_Footer_Sig[] = { '_', '_', 'S', 'B', 'F', 'T', 'R', '\0' };
+            if (m_codeSize > (sizeof(SB_Footer_Sig) + sizeof(DWORD)))
+            {
+                auto ptr = reinterpret_cast<BYTE*>(BaseIP) + length; // data end
+                ptr -= sizeof(SB_Footer_Sig);
+                if (memcmp(ptr, &SB_Footer_Sig, sizeof(SB_Footer_Sig)) == 0)
+                {
+                    ptr -= sizeof(DWORD);
+                    auto codeSize = MemRead<DWORD>(ptr);
+
+                    if (codeSize < m_codeSize) m_codeSize = codeSize;
+                }
+            }
+
             m_codeChecksum = crc32(reinterpret_cast<BYTE*>(BaseIP), length);
 
             // thread name from filename
@@ -370,6 +386,18 @@ std::string CCustomScript::GetInfoStr(bool currLineInfo) const
             else
             {
                 ss << ": ...";
+            }
+
+            // add previously executed command info if available
+            if (CCustomOpcodeSystem::prevOpcode != 0xFFFF)
+            {
+                ss << " \nPreviously called command: ";
+
+                auto commandName = CleoInstance.OpcodeInfoDb.GetCommandName(CCustomOpcodeSystem::prevOpcode);
+                if (commandName)
+                    ss << commandName;
+                else
+                    ss << "[" << std::hex << std::uppercase << std::setw(4) << std::setfill('0') << CCustomOpcodeSystem::prevOpcode << "]";
             }
         }
     }